ADD params and comments to the bin scripts

+++ bin/stop.sh
@@ -1,4 +1,33 @@
 #!/bin/bash
+# This script stops docker container(s) and unused networks
+# If PROJECT_NAME doesn't exists then script will stop all exists containers and unused networks
+#
+# Format: ./stop.sh [PARAMS] [-n PROJECT_NAME]
+# Params:
+#   -c - remove containers after they stops
+#   -a - remove all containers and their images after they stops
+
+# get params
+while [[ $# -gt 0 ]]
+do
+    key="$1"
+    case $key in
+        -c) isRemoveContainers='true';;
+        -a) isRemoveAll='true';;
+        -n)
+            if [ ! -z "$2" ]; then
+                export PROJECT="$2"
+            fi
+            shift
+            ;;
+        *)
+            echo "Invalid option -$1"
+            exit
+            ;;
+    esac
+        shift
+done
+
 
 # bin dir
 DC_BIN_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
@@ -9,42 +38,52 @@ DC_PROJECT_DIR="$DC_ROOT_DIR/projects"
 
 
 
-PROJECT=$1
-cd "$DC_ROOT_DIR"
+# one/all processing
 if [ ! -z "$PROJECT" ]
 then
-    # set colors
-#    GREEN='\033[0;32m'
-#    NC='\033[0m' # No Color
-    # stop container
-#    echo -n "Stopping $PROJECT ... "
-#    docker stop "$PROJECT" > /dev/null 2>&1
-#    echo -e "${GREEN}done${NC}"
-
-    # remove container
-#    echo -n "Removing $PROJECT ... "
-#    docker rm -f "$PROJECT" > /dev/null 2>&1
-#    echo -e "${GREEN}done${NC}"
-
-    # docker compose down
-    cd "$DC_PROJECT_DIR/$PROJECT"
-#    docker-compose down -rmi all > /dev/null 2>&1
-#    docker-compose down --rmi all
-    docker-compose down
+    # if project exists
+    if [ -d "$DC_PROJECT_DIR/$PROJECT" ]
+    then
+        cd "$DC_PROJECT_DIR/$PROJECT"
+        # remove all
+        if [ ! -z "$isRemoveAll" ]
+        then
+            docker-compose down --rmi all
+        # remove containers
+        elif [ ! -z "$isRemoveContainers" ]
+        then
+            docker-compose down
+        # only stop containers
+        else
+            docker-compose stop
+        fi
+    else
+        echo "Invalid project - $PROJECT"
+    fi
 else
     cd "$DC_ROOT_DIR"
+    CONTAINERS=$(docker ps -a -q)
 
     # stop all containers
     echo "Stopped containers (id):"
-    docker stop $(docker ps -a -q)
+    docker stop $CONTAINERS
 
     # remove all containers
-    echo "Removed containers (id):"
-    docker rm -f $(docker ps -a -q)
+    if [ ! -z "$isRemoveAll" ] || [ ! -z "$isRemoveContainers" ]
+    then
+        echo "Removed containers (id):"
+        docker rm -f $CONTAINERS
+    fi
 
-    #docker rmi $(docker images -q)
-
-    # remove all unused networks
-    #docker network rm $(docker network list -q)
-    docker network prune -f
+    # remove all images
+    if [ ! -z "$isRemoveAll" ]
+    then
+        echo "Removed images (id):"
+        docker rmi $(docker images -q)
+    fi
 fi;
+
+
+
+# remove all unused networks
+docker network prune -f
